+++ packages/examples/src/creating_a_ui/model_data.ts
@@ -7,64 +7,78 @@ export async function createData(manager: ModelManager) {
     // Create Users
 
     const joe = (await manager.create(new User({
-        username: 'joe123'
+        username: 'joe123',
+        first_name: 'Joe',
+        last_name: 'Mahony'
     }))).result;
 
     const bill = (await manager.create(new User({
-        username: 'bill27'
+        username: 'bill27',
+        first_name: 'Bill',
+        last_name: 'Biddington'
     }))).result;
 
     const jane = (await manager.create(new User({
-        username: 'jane12'
+        username: 'jane12',
+        first_name: 'Jane',
+        last_name: 'Foster'
     }))).result;
 
     // Create Posts
 
     const post1 = (await manager.create(new Post({
+        post_date: '2018-03-18',
         title: 'Programmer',
-        body: 'Nader-Monahan',
+        description: 'Nader-Monahan, Rebum brute est.',
         user: jane
     }))).result;
 
     const post2 = (await manager.create(new Post({
+        post_date: '2018-02-23',
         title: 'Design Engineer',
-        body: 'Hyatt Group',
+        description: 'Hyatt Group, vel utroque constituam.',
         user: bill
     }))).result;
 
     await manager.create(new Post({
+        post_date: '2017-11-08',
         title: 'VP Sales',
-        body: 'Senger-Hoeger',
+        description: 'Senger-Hoeger, ad quas patrioque eloquentiam.',
         user: jane
     }));
 
     await manager.create(new Post({
+        post_date: '2018-01-01',
         title: 'VP Operations',
-        body: 'Schamberger-Orn',
+        description: 'Schamberger-Orn, ea malis graeco copiosae.',
         user: joe
     }));
 
     await manager.create(new Post({
+        post_date: '2018-01-15',
         title: 'Administrative Assistant',
-        body: 'Hammes-Dooley',
+        description: 'Hammes-Dooley, eum in labitur dignissim.',
         user: jane
     }));
 
     await manager.create(new Post({
+        post_date: '2018-02-03',
         title: 'Tax Accountant',
-        body: 'Corkery, Wyman and Brekke',
+        description: 'Corkery, Wyman and Brekke',
         user: bill
     }));
 
     await manager.create(new Post({
+        post_date: '2017-12-20',
         title: 'Food Chemist',
-        body: 'Friesen-Reichel',
+        description: 'Friesen-Reichel, modo eius ut eos.',
         user: joe
     }));
 
     await manager.create(new Post({
+        post_date: '2018-03-05',
         title: 'GIS Technical Architect',
-        body: 'Hyatt-Goyette',
+        description: 'Hyatt-Goyette, utamur corrumpit ea sed.',
         user: joe
     }));